wdobackend/requirements/base.txt
cachetools==5.3.1
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.10.8
django-allauth==0.33.0
django-cors-headers==2.1.0
django-rest-auth==0.9.1
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
facebook-sdk==2.0.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
idna==2.6
oauthlib==2.0.3
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==1.5.3
PyMySQL==1.1.0
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
rsa==4.9
six==1.10.0
urllib3==1.22
